chore: refactor components
this commit splits components.js into individual files (by purpose,
hopefully) and update all imports. also extracts more inline code.

i think at some point, we should reach for an actual templating
library and maybe a build tool like vite to have nicer imports than
now.
